move *Info classes to web package
