+++ service-team/src/main/java/sg/ncl/service/team/data/jpa/TeamEntity.java
@@ -3,7 +3,7 @@ package sg.ncl.service.team.data.jpa;
 import org.hibernate.annotations.GenericGenerator;
 import sg.ncl.common.jpa.AbstractEntity;
 import sg.ncl.service.team.domain.*;
-import sg.ncl.service.team.dtos.TeamMemberInfo;
+import sg.ncl.service.team.web.TeamMemberInfo;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
