+++ service-team/src/test/java/sg/ncl/service/team/Util.java
@@ -4,7 +4,7 @@ import org.apache.commons.lang3.RandomStringUtils;
 import sg.ncl.service.team.data.jpa.TeamEntity;
 import sg.ncl.service.team.data.jpa.TeamMemberEntity;
 import sg.ncl.service.team.domain.TeamMemberType;
-import sg.ncl.service.team.dtos.TeamMemberInfo;
+import sg.ncl.service.team.web.TeamMemberInfo;
 
 import java.time.ZonedDateTime;
